LayoutTests/fast/frames/seamless/seamless-float-expected.txt

Test that floated seamless iframes 'shrink-wrap' their contents, as floated divs would.
PASS window.getComputedStyle(iframe1).width is "150px"
FAIL window.getComputedStyle(iframe1).height should be 50px. Was 100px.
PASS window.getComputedStyle(iframe2).width is "100px"
PASS window.getComputedStyle(iframe2).height is "100px"
